% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{hline}
\alias{hline}
\title{dotted line bisecting y axis}
\usage{
hline(y = 0, colour = "#4d4d4d")
}
\arguments{
\item{y}{coordinate to place line}

\item{colour}{colour of line}
}
\value{
line to add to plotly
}
\description{
dotted line bisecting y axis
}
\examples{
hline(0)
}
