% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rand_coord_opt2}
\alias{rand_coord_opt2}
\title{returns co-ordinates for point labels while trying to spread points from eachother and x and y axis}
\usage{
rand_coord_opt2(n, xmin, xmax, ymin, ymax, min_distance, random_state = 12)
}
\arguments{
\item{n}{number of points requiring label}

\item{xmin}{min x-coord}

\item{xmax}{maxx x-coord}

\item{ymin}{min y-coord}

\item{ymax}{max y-coord}

\item{min_distance}{miin distance between labels}

\item{random_state}{random state}
}
\value{
poistions for labels
}
\description{
returns co-ordinates for point labels while trying to spread points from eachother and x and y axis
}
\examples{
rand_coord_opt2(n = 1, 
xmin = 40, 
xmax = 100, 
ymin = 40, 
ymax = 100, 
min_distance = 2, 
random_state = 12)
}
