% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Call function to launch app}
\usage{
run_app()
}
\value{
launches app in which you can upload data exported from google keywords for analysis
}
\description{
Call function to launch app
}
\examples{
run_app()
}
